+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2024-07-03
+
+### Added
+- Added the IRA page with the EQIP section to display 2023 data and predicted 2024-2031 data with total dollar amount and practice options [#268](https://github.com/policy-design-lab/pdl-frontend/issues/268) 
+
 ## [0.20.0] - 2024-06-03
 
 ### Changed
@@ -241,6 +246,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Map data json [#12](https://github.com/policy-design-lab/pdl-frontend/issues/12)
 - Final landing page changes for initial milestone [#15](https://github.com/policy-design-lab/pdl-frontend/issues/15)
 
+[1.0.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.20.0...1.0.0
 [0.20.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.19.0...0.20.0
 [0.19.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.18.0...0.19.0
 [0.18.0]: https://github.com/policy-design-lab/pdl-frontend/compare/0.17.0...0.18.0